fix: address PR review feedback
- Extract parseUserIdRefs() helper into refs.ts to deduplicate user ID
  parsing between reply and create commands
- Fix process.exit mock in test to throw so execution halts properly
- Add --dry-run to thread create skill content documentation
